+++ src/tsp/io/point.rs
@@ -1,21 +1,21 @@
 #[derive(Debug)]
 pub struct Point {
-    x: isize,
-    y: isize
+    x: i64,
+    y: i64
 }
 
 impl Point {
-    pub fn new(x: isize, y: isize) -> Self{
+    pub fn new(x: i64, y: i64) -> Self{
         Point {
             x,
             y
         }
     }
 
-    pub fn calculate_distance(&self, point: &Point) -> isize {  // Calculates the cartesian distance between two points
+    pub fn calculate_distance(&self, point: &Point) -> i64 {  // Calculates the cartesian distance between two points
         let a = ((self.x - point.x) as f64).powf(2f64);
         let b = ((self.y - point.y) as f64).powf(2f64);
-        let dist = (a + b).sqrt().round() as isize;
+        let dist = (a + b).sqrt().round() as i64;
         dist
     }
 }
